Refactor request handling and error management: Update async functions in `test.sh` to ignore request parameters, introduce `ApiError` for better error handling in `api_error.rs`, and restructure request body handling in `request_body.rs`. Clean up imports and enhance type definitions in `types.rs` and `mod.rs` for improved clarity and organization.
